Pass [10] complete: componentization, guest-world synthesis, artifact set
Milestone 1 step 8. The core module is wrapped as a Component Model
component: the compiler synthesizes the guest's mirror world — importing
exactly the clean:host interfaces the program calls (uncalled imports are
pruned at MIR with index remapping) and exporting the entry points the
program defines — embeds its metadata, and componentizes with validation.
The `handle` export gets a boundary shim (world says handler-id: u32 →
core i32; Clean's integer is i64). Assembly failures surface as COM013
with the Platform 10 template (CMP-04).

compile() now returns the full artifact set: component.wasm,
build-manifest.json (§14.8; compiler.sha256 and zeroed timings are
recorded M1 placeholders — the timings/CMP-02 tension goes to the brief's
Discoveries), diagnostics, exit-0 CLI writes.

Acceptance: the full 9a guest — 5 routes, enum/string/record register,
option+default, bytes echo, constant field list — compiles to a component
that `wasm-tools component targets` validates against the guest world
(Discovery: the brief's `targets -w server` reading cannot hold for a
guest; `server` is the host's world — the guest targets the mirror world
exactly as clean-server's fake-guest does), and all three artifacts are
byte-deterministic (CMP-02).

+++ crates/clean-compiler/tests/canonical_abi.rs
@@ -349,7 +349,7 @@ functions:
         .instantiate(&mut store, &module)
         .expect("instantiates");
     let handle = instance
-        .get_typed_func::<i64, ()>(&mut store, "handle")
+        .get_typed_func::<i32, ()>(&mut store, "handle")
         .expect("handle export");
     handle.call(&mut store, 1).expect("handle runs");
     assert_eq!(store.data().log.as_slice(), ["42".to_string()]);
@@ -366,7 +366,7 @@ functions:
         .instantiate(&mut store, &module)
         .expect("instantiates");
     let handle = instance
-        .get_typed_func::<i64, ()>(&mut store, "handle")
+        .get_typed_func::<i32, ()>(&mut store, "handle")
         .expect("handle export");
     handle.call(&mut store, 1).expect("handle runs");
     assert_eq!(store.data().log.as_slice(), ["no id".to_string()]);
@@ -447,7 +447,7 @@ functions:
         .instantiate(&mut store, &module)
         .expect("instantiates");
     instance
-        .get_typed_func::<i64, ()>(&mut store, "handle")
+        .get_typed_func::<i32, ()>(&mut store, "handle")
         .expect("handle export")
         .call(&mut store, 4)
         .expect("handle runs");
@@ -523,7 +523,7 @@ functions:
         .instantiate(&mut store, &module)
         .expect("instantiates");
     instance
-        .get_typed_func::<i64, ()>(&mut store, "handle")
+        .get_typed_func::<i32, ()>(&mut store, "handle")
         .expect("handle export")
         .call(&mut store, 6)
         .expect("handle runs");
